requirements.txt
streamlit
google-auth-oauthlib
google-api-python-client
pandas
requests
